Fix fingerprint match failure in trident devices
The bug was due to passing a wrong bitmap fingerprint image to the
convertToFmdAndMatch method. The issue was present only in trident
devices due to the fact that the wrong bitmap was passed only when
fingerprint is captured using OnFingerprintGrabbedFullListener.
This usage is only in trident devices.

The fix takes into account which type of fingerprint is being compared
and passes the correct fingerprint bitmap to the convertToFmdAndMatch
method. If the scan type is two fingers split, then only first
fingerprint bitmap is compared.